+++ src/app/admin-panel/deposit-place/deposit-place.component.ts
@@ -59,14 +59,16 @@ export class DepositPlaceComponent implements OnInit {
     const index = this.depositPlaces.indexOf(place);
 
     if (index !== -1) {
-      this.loading = true;
-      this.crudService.deleteById(this.route + place._id).subscribe(response => {
-        this.depositPlaces.splice(index, 1);
-        this.loading = false;
-      }, error => {
-        window.alert(error);
-        this.loading = false;
-      });
+      if (window.confirm('Você tem certeza?')) {
+        this.loading = true;
+        this.crudService.deleteById(this.route + place._id).subscribe(response => {
+          this.depositPlaces.splice(index, 1);
+          this.loading = false;
+        }, error => {
+          window.alert(error);
+          this.loading = false;
+        });
+      }
     }
   }
 
@@ -76,6 +78,7 @@ export class DepositPlaceComponent implements OnInit {
     }, error => {
       window.alert(error);
     });
+
   }
 
   ngOnInit() {
